GDC_dna_seq: fixed Start widget extension parsing
input file name extension parsing failed if there was a period in the filename,
this update will check against a known list of extensions but allow users to
input a custom extension also.
Updated biobambam to allow this widget to be bypassed if a fastq file is used
instead of a bam file.

Signed-off-by: Bob Schmitz III <14095796+rgschmitz1@users.noreply.github.com>